.tests/requirements.txt
pymongo
pytest
requests
paho-mqtt
